Initial add of the IPL procs index files
These files contain summaries of the documentation of the IPL proc files.

The descriptions file is a summary of the header and documentation comments
for each of the IPL files.
The procs file is a summary of each procedure/record/global/link/invocable/class
declaration within the relavant IPL file.

An example of the description of a procedure contains the paramters and it
expected type and the return types/suspend types/fail/error for that
procedure.

Signed-off-by: Bruce Rennie <brennie@dcsi.net.au>